* base: Don't allow negative levels in GridGenerators. * optimization: Remove redundant specification of dimension in Interpolant* constructors.
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@4247 4eea3252-f0fb-4393-894d-40516dce545b

+++ base/src/sgpp/base/grid/generation/StandardGridGenerator.hpp
@@ -32,9 +32,9 @@ namespace SGPP {
          */
         virtual ~StandardGridGenerator();
 
-        virtual void regular(int level);
-        virtual void cliques(int level, size_t clique_size);
-        virtual void full(int level);
+        virtual void regular(size_t level);
+        virtual void cliques(size_t level, size_t clique_size);
+        virtual void full(size_t level);
         virtual void refine(RefinementFunctor* func);
         virtual size_t getNumberOfRefinablePoints();
 
@@ -42,8 +42,8 @@ namespace SGPP {
         virtual void coarsenNFirstOnly(CoarseningFunctor* func, DataVector* alpha, size_t numFirstOnly);
         virtual size_t getNumberOfRemovablePoints();
 
-        virtual void refineMaxLevel(RefinementFunctor* func, int maxLevel);
-        virtual size_t getNumberOfRefinablePointsToMaxLevel(int maxLevel);
+        virtual void refineMaxLevel(RefinementFunctor* func, size_t maxLevel);
+        virtual size_t getNumberOfRefinablePointsToMaxLevel(size_t maxLevel);
 
       protected:
         /// pointer to the storage object
